introduce new contact class for API resource with valid addr1 and addr2 json properties. fix unit tests names.
