+++ MailChimp.Net.Tests/ApiTest.cs
@@ -1,35 +1,31 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ApiTest.cs" company="Brandon Seydel">
+// <copyright file="AuthorizedAppTest.cs" company="Brandon Seydel">
 //   N/A
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
 
 using System.Threading.Tasks;
-
+using MailChimp.Net.Models;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace MailChimp.Net.Tests
 {
     /// <summary>
-    /// The authorized app test.
+    /// The api test.
     /// </summary>
     [TestClass]
-    public class AuthorizedAppTest : MailChimpTest
+    public class ApiTest : MailChimpTest
     {
         /// <summary>
-        /// The should_ return_ app_ information.
+        /// The should_ return_ ap i_ information.
         /// </summary>
         /// <returns>
         /// The <see cref="Task"/>.
         /// </returns>
         [TestMethod]
-        public async Task Should_Return_App_Information()
+        public async Task Should_Return_API_Information()
         {
-            var apiInfo = await this._mailChimpManager.Apps.GetAllAsync();
-            apiInfo = await this._mailChimpManager.Configure(new MailChimpConfiguration
-            {
-                Limit = 100000
-            }).Apps.GetAllAsync();
+            var apiInfo = await this._mailChimpManager.Api.GetInfoAsync();
             Assert.IsNotNull(apiInfo);
         }
     }
