+++ MailChimp.Net.Tests/AuthorizedAppTest.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="AuthorizedAppTest.cs" company="Brandon Seydel">
+// <copyright file="ApiTest.cs" company="Brandon Seydel">
 //   N/A
 // </copyright>
 // --------------------------------------------------------------------------------------------------------------------
@@ -11,21 +11,25 @@ using Microsoft.VisualStudio.TestTools.UnitTesting;
 namespace MailChimp.Net.Tests
 {
     /// <summary>
-    /// The api test.
+    /// The authorized app test.
     /// </summary>
     [TestClass]
-    public class ApiTest : MailChimpTest
+    public class AuthorizedAppTest : MailChimpTest
     {
         /// <summary>
-        /// The should_ return_ ap i_ information.
+        /// The should_ return_ app_ information.
         /// </summary>
         /// <returns>
         /// The <see cref="Task"/>.
         /// </returns>
         [TestMethod]
-        public async Task Should_Return_API_Information()
+        public async Task Should_Return_App_Information()
         {
-            var apiInfo = await this._mailChimpManager.Api.GetInfoAsync();
+            var apiInfo = await this._mailChimpManager.Apps.GetAllAsync();
+            apiInfo = await this._mailChimpManager.Configure(new MailChimpConfiguration
+            {
+                Limit = 100000
+            }).Apps.GetAllAsync();
             Assert.IsNotNull(apiInfo);
         }
     }
