+++ MailChimp.Net/MailChimp.Net.csproj
@@ -118,6 +118,7 @@
     <Compile Include="Interfaces\IECommerceCartLogic.cs" />
     <Compile Include="Logic\ECommerceProductVarianceLogic.cs" />
     <Compile Include="Models\Address.cs" />
+    <Compile Include="Models\ApiContact.cs" />
     <Compile Include="Models\AutomationStatus.cs" />
     <Compile Include="Models\BatchDelivery.cs" />
     <Compile Include="Models\CampaignAction.cs" />
